requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
pytest==8.2.0
httpx==0.27.0
openai==1.23.2
APScheduler==3.10.4
# Data & persistence
sqlmodel==0.0.16
asyncpg==0.29.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
loguru==0.7.2
anyio==4.3.0
